Attempting to fix integrity error on updateOrCreate()
Issue is likely caused by Laravel race condition bug.
https://github.com/laravel/framework/issues/19372

According to the bug report it's unlikely to be fixed in laravel,
as such I have switched to doing manual updating + creating without
using the function to try and avoid the annoying integrity constraint
problem that is happening when it shouldn't. Yes Laravel the ID is a dup,
that is why I am telling you to UPDATEorCreate *facepalm*.